pyproject.toml
[project]
name = "goit-pythonweb-hw-12"
version = "0.1.0"
description = ""
authors = [
    {name = "Dayout",email = "fearess7@gmail.com"}
]

[tool.poetry.dependencies]
python = "^3.11"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0"
alembic = "^1.14.0"
greenlet = "^3.1.1"
fastapi = ">=0.100.0"
dotenv = "^0.9.9"
email-validator = "^2.2.0"
uvicorn = "^0.34.1"
pydantic = ">=2.0"
slowapi = "^0.1.9"
pydantic-settings = "^2.8.1"
fastapi-mail = "^1.4.2"
cloudinary = "^1.44.0"
python-jose = "^3.4.0"
libgravatar = "^1.0.4"
redis = "^5.2.1"
redis-lru = "^0.1.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2-binary = "^2.9.10"
python-multipart = "^0.0.20"
bcrypt = "<4.0"
aiocache = "^0.12.3"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"
pytest-cov = "^6.1.1"
aiosqlite = "^0.21.0"

[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.3"

[tool.poetry]
packages = [{ include = "src" }]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
